+++ src/MudBlazor/Utilities/Converter/DefaultConverter.Number.cs
@@ -1,0 +1,44 @@
+﻿// Copyright (c) MudBlazor 2021
+// MudBlazor licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Globalization;
+using System.Numerics;
+using MudBlazor.Resources;
+using MudBlazor.Utilities.Converter.Base;
+
+namespace MudBlazor.Utilities.Converter;
+
+#nullable enable
+internal partial class DefaultConverter
+{
+    internal sealed class NumberConverter<TNumber>(Func<CultureInfo> culture, Func<string?> format)
+        : IReversibleConverter<TNumber, string?>
+        where TNumber : INumber<TNumber>
+    {
+        public string Convert(TNumber input)
+        {
+            var currentCulture = culture.Invoke();
+            var currentFormat = format.Invoke();
+
+            return input.ToString(currentFormat, currentCulture);
+        }
+
+        public TNumber ConvertBack(string? input)
+        {
+            if (string.IsNullOrEmpty(input))
+            {
+                return TNumber.Zero;
+            }
+
+            var currentCulture = culture.Invoke();
+
+            if (TNumber.TryParse(input, NumberStyles.Any, currentCulture, out var result))
+            {
+                return result;
+            }
+
+            throw new ConversionException(LanguageResource.Converter_InvalidNumber);
+        }
+    }
+}
